requirements.txt
django==4.1.4
tabdeal-python==0.4.6
djangorestframework==3.14.0